+++ example/lib/main.dart
@@ -33,6 +33,7 @@ import 'package:z_components/components/confirmacao-de-previsto/atualizar-dados-
 import 'package:z_components/components/z-expendable-item-tile.dart';
 import 'package:z_components/view-model/buscarinfo-viewmodel.dart';
 import 'package:z_components/components/utils/dialog-utils.dart';
+import 'package:z_components/components/confirmar-senha/confirmar-senha.dart';
 
 
 import 'main-testing.dart';
@@ -54,8 +55,7 @@ class ComponentExemploClasse extends StatefulWidget {
   _ComponentExemploClasseState createState() => _ComponentExemploClasseState();
 }
 
-class _ComponentExemploClasseState extends State<ComponentExemploClasse>
-    with AfterLayoutMixin<ComponentExemploClasse> {
+class _ComponentExemploClasseState extends State<ComponentExemploClasse> {
   var controllerEmail = new TextEditingController();
   var controlerPadrao = new TextEditingController();
   var controllerNome = new TextEditingController();
@@ -237,9 +237,10 @@ List<String> orem = ["ENTRADA","SAÍDA","ENTRADA","SAÍDA","ENTRADA","SAÍDA"];
     return new Scaffold(
       backgroundColor: Colors.black,
         floatingActionButton: FloatingActionButton(
-          onPressed: () {
-            _dialogUtils.showAlertDialogErro("Sem conexão com a internet",
-                "Não é possivel esta funcionalidade sem acesso à internet");
+          onPressed: () async {
+         ConfirmarSenha confirmar = new ConfirmarSenha(token: "eyJhbGciOiJSUzI1NiIsImtpZCI6ImY4NzJlODA0NjVjMzk1NGY2MTE5YWNmZTRiYzc1NDQ4IiwidHlwIjoiSldUIn0.eyJuYmYiOjE1ODk4MjcxODMsImV4cCI6MTU4OTgzMDc4MywiaXNzIjoiaHR0cHM6Ly9pZGVudGl0eS1zZXJ2ZXItZGV2LnplbGxhci5jb20uYnIiLCJhdWQiOlsiaHR0cHM6Ly9pZGVudGl0eS1zZXJ2ZXItZGV2LnplbGxhci5jb20uYnIvcmVzb3VyY2VzIiwibW9sdHJlcy5hY2Vzc28uYXBpIl0sImNsaWVudF9pZCI6IlpDb2xhYm9yYWRvciIsInN1YiI6IjJiZmFhY2I2LWVkMjYtNDFiNC1iZTM4LWUyMzI5Yjk0MmM1YSIsImF1dGhfdGltZSI6MTU4OTgyNzE4MywiaWRwIjoibG9jYWwiLCJBc3BOZXQuSWRlbnRpdHkuU2VjdXJpdHlTdGFtcCI6IkFJUExGNkxBSUU2Q01ON01ONjRFUE1HS0xWRlpIWkFRIiwiYWNjb3VudCI6IlplbGxhcjIiLCJpZEFjY291bnQiOiI4RTZFQjYzMi1CNzA3LTQxM0YtQTE1Mi0zQ0ZGRDFGOTgyQjUiLCJpZENvbGFib3JhZG9yIjoiMjZBQTMxMkYtQ0Q3OS00N0M5LUE0RjItQkI0RkVCNzM3MjkyIiwicHJlZmVycmVkX3VzZXJuYW1lIjoiMzk5LjEwOC44NDgtNzkiLCJlbWFpbCI6InZpY3RvcnNvdXp6QGdtYWlsLmNvbSIsImVtYWlsX3ZlcmlmaWVkIjpmYWxzZSwicGhvbmVfbnVtYmVyX3ZlcmlmaWVkIjpmYWxzZSwibmFtZSI6IlZpY3RvciBEZSBTb3V6YSBBbHZlcyIsInBob25lX251bWJlciI6IigxMSkgOTg0NzAtMzMzNiIsInNjb3BlIjpbIm9wZW5pZCIsInByb2ZpbGUiLCJlbWFpbCIsIm1vbHRyZXMuYWNlc3NvLmFwaS5mdWxsIiwib2ZmbGluZV9hY2Nlc3MiXSwiYW1yIjpbInB3ZCJdfQ.R9hxcChbo3PgX8KFQ8Ml0F2FeTE6r83vnBUFRNMPhHFCU0k0AasUmhoL9pDSErSWk32YVi9RXJ9PLAql8-Q4ON1YYfKvzBUvhQEotpDS1OUwLmwoD46PdU80FQ3niTnQEz5--RTYJ0-C6UZyC26nXuTHp12vMm60C3McAVcCbvlxAHlI6aXllgvxD3rMkvj9Xm8C8d5IeavdxxtJvN67sVsefj2o7aNsvDB_kb0D-7exufW5pphO1SOHL1DpnaSY309nu1mEbHeqJeX8ardC54vs2qJRDbe1d0Cx5g6XFKIQjtaJ_WzVR6O4gALkasbxelsdfyyya6P66qFVUbxGMA", context: context);
+        bool teste = await  confirmar.dialogConfirmarSenha();
+        print(teste);
           },
         ),
         appBar: new AppBar(),