+++ lib/api/confirmar-senha/i-confirmar-senha-service.dart
@@ -1,0 +1,4 @@
+
+abstract class IConfirmarSenhaService{
+  Future<bool> validarSenha(String pin);
+}